core app funtionality complete

+++ client/src/components/TodoList.js
@@ -2,6 +2,9 @@ import React, { useState, useEffect } from 'react';
 
 function TodoList({ username, refresh }) {
   const [todos, setTodos] = useState([]);
+  const [editMode, setEditMode] = useState(false);
+  const [editedTaskName, setEditedTaskName] = useState('');
+  const [editTodoId, setEditTodoId] = useState(null);
   
  username = localStorage.getItem('username');
 
@@ -44,7 +47,11 @@ useEffect(() => {
 }, [username]);
 
   
-  const handleDeleteTodo = async (todoId) => {
+const handleDeleteTodo = async (todoId) => {
+  const shouldDelete = window.confirm('Are you sure you want to delete this to-do?');
+  if (!shouldDelete) {
+    return;
+  }
     try {
       const response = await fetch(`/deleteTodo/${todoId}`, {
         method: 'DELETE',
@@ -61,12 +68,71 @@ useEffect(() => {
   };
 
   
+  const handleEditTodo = (todoId, taskName) => {
+    setEditMode(true);
+    setEditTodoId(todoId);
+    setEditedTaskName(taskName);
+  };
+
+  const handleCancelEdit = () => {
+    setEditMode(false);
+    setEditTodoId(null);
+    setEditedTaskName('');
+  };
+
+  const handleSaveEdit = async () => {
+    try {
+      const response = await fetch(`/editTodos/${editTodoId}`, {
+        method: 'PUT',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({ task_name: editedTaskName }),
+      });
+
+      if (!response.ok) {
+        throw new Error('Network response was not ok');
+      }
+
+      // Update the todos list after successful edit
+      const updatedTodos = [...todos];
+      const editedTodoIndex = updatedTodos.findIndex((todo) => todo.id === editTodoId);
+      if (editedTodoIndex !== -1) {
+        updatedTodos[editedTodoIndex].task_name = editedTaskName;
+        setTodos(updatedTodos);
+      }
+
+      // Reset edit mode
+      handleCancelEdit();
+    } catch (error) {
+      console.error('Error editing todo:', error);
+    }
+  };
+  
   return (
     <div>
       <h2>Todo List</h2>
       <ul>
         {todos.map((todo) => (
-          <li key={todo.id}>{todo.task_name}<button onClick={() => handleDeleteTodo(todo.id)}>Delete</button></li>
+          <li key={todo.id}>
+            {editMode && editTodoId === todo.id ? (
+              <>
+                <input
+                  type="text"
+                  value={editedTaskName}
+                  onChange={(e) => setEditedTaskName(e.target.value)}
+                />
+                <button onClick={handleSaveEdit}>Save</button>
+                <button onClick={handleCancelEdit}>Cancel</button>
+              </>
+            ) : (
+              <>
+                <button onClick={() => handleEditTodo(todo.id, todo.task_name)}>Edit</button>
+                {todo.task_name}
+                <button onClick={() => handleDeleteTodo(todo.id)}>Delete</button>
+              </>
+            )}
+          </li>
         ))}
       </ul>
     </div>
